+++ src/DataWrangling/DataWrangling.jl
@@ -73,7 +73,7 @@ end
 
 include("metadata.jl")
 include("inpaint_mask.jl")
-include("JRA55.jl")
+include("JRA55/JRA55.jl")
 include("ECCO/ECCO.jl")
 
 using .ECCO
